Update tests and support refreshing on collection elements
