+++ src-shared/CollectionElement.ts
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
 export interface CollectionElement {
     id?: string;
     value: string;
