+++ src-shared/CollectionWebviewPayload.ts
@@ -1,6 +1,10 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
 import { CollectionElement } from './CollectionElement';
 
 export interface CollectionWebviewPayload {
     data: CollectionElement[];
+    clearCache: boolean;
     hasMore: boolean;
 }
