+++ src-webview/data-viewer/CollectionType.ts
@@ -1,1 +1,4 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
 export type CollectionType = 'list' | 'set' | 'zset' | 'hash';
