+++ src-webview/data-viewer/CollectionView.css
@@ -1,3 +1,6 @@
+/* Copyright (c) Microsoft Corporation.
+   Licensed under the MIT License. */
+
 .container {
     height: 100vh;
     display: flex;
