+++ src-webview/data-viewer/SelectableCollectionElement.ts
@@ -1,7 +1,5 @@
-export interface CollectionElement {
-    id?: string;
-    value: string;
-}
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
 
 export interface SelectableCollectionElement {
     id?: string;
