--- src-webview/data-viewer/Split.ts
@@ -1,2 +1,0 @@
-declare module 'react-split';
-declare module 'react-resize-panel';
